payment noti 2

+++ src/main/java/com/tanmeyah/postoffice/DTO/Projection/InquiryProjection.java
@@ -1,9 +1,11 @@
 package com.tanmeyah.postoffice.DTO.Projection;
 
+import java.math.BigDecimal;
+
 public interface InquiryProjection {
     String getIdno();
     String getMobilephone();
     String getFullName();
     Integer getDrc();
-    Double getTotalAmt();
+    BigDecimal getTotalAmt();
 }
